osm2ed: import car speed in ed database
This PR import car speed on every edges of the routing graph.
The tag `maxspeed` is used if present else a speed is chosen from the
highway type.
A factor is applied on max speed to get a somewhat realistic speed for
the edge.

Default value are valid for France and should be configurable if we want
to use it internationally. Rural zone vs urban zone isn't handled so
secondary and tertiary highway have a max speed of 50km/h everywhere if
not specified by a tag, primary are at 80km/h so sometime max speed is
higher that the legal value :/